+++ src/popgenIO/Formats/FlatFile.java
@@ -79,7 +79,7 @@ public class FlatFile {
 				String[] cols = fst.nextLine().trim().split("\\s");
 				double position = t;
 				String name = null;
-				int[] alleles = null;
+				byte[] alleles = null;
 				
 				if (cols.length >= 1) {
 					position = Double.parseDouble(cols[0]);
@@ -91,7 +91,7 @@ public class FlatFile {
 
 				if (cols.length >= 3) {
 					assert cols[2].length() == 3;
-					alleles = new int[] { Character.getNumericValue(cols[2].charAt(0)), Character.getNumericValue(cols[2].charAt(2)) };
+					alleles = new byte[] { (byte) Character.getNumericValue(cols[2].charAt(0)), (byte) Character.getNumericValue(cols[2].charAt(2)) };
 				}
 				data.addSite(new Site(t, position, name, alleles).globalize());
 			}
